Moving map init to its own files
