+++ ai_animation/assets/maps/standard/styles.json
@@ -5,9 +5,6 @@
   "impassable": {
     "fill": "#353433"
   },
-  "background": {
-    "fill": "#FFFFFF"
-  },
   "labeltext": {
     "font-size": "12px",
     "fill": "black",
@@ -36,31 +33,6 @@
     "fill-opacity": "0.0",
     "opacity": "0.0"
   },
-  "provinceRed": {
-    "fill": "url(#patternRed)",
-    "stroke": "black",
-    "stroke-width": "2"
-  },
-  "provinceBrown": {
-    "fill": "url(#patternBrown)",
-    "stroke": "black",
-    "stroke-width": "2"
-  },
-  "provinceGreen": {
-    "fill": "url(#patternGreen)",
-    "stroke": "black",
-    "stroke-width": "2"
-  },
-  "provinceBlack": {
-    "fill": "url(#patternBlack)",
-    "stroke": "black",
-    "stroke-width": "2"
-  },
-  "provinceBlue": {
-    "fill": "url(#patternBlue)",
-    "stroke": "black",
-    "stroke-width": "2"
-  },
   "nopower": {
     "fill": "#ddd2af",
     "stroke": "black",
@@ -72,131 +44,5 @@
     "stroke": "black",
     "stroke-linejoin": "round",
     "stroke-width": "2"
-  },
-  "britain": {
-    "fill": "royalblue",
-    "stroke": "black",
-    "stroke-width": "2"
-  },
-  "egypt": {
-    "fill": "#808000",
-    "stroke": "black",
-    "stroke-width": "2"
-  },
-  "france": {
-    "fill": "#00FFFF",
-    "stroke": "black",
-    "stroke-width": "2"
-  },
-  "germany": {
-    "fill": "darkgrey",
-    "stroke": "black",
-    "stroke-width": "2"
-  },
-  "italy": {
-    "fill": "#80FF80",
-    "stroke": "black",
-    "stroke-width": "2"
-  },
-  "poland": {
-    "fill": "#FF0000",
-    "stroke": "black",
-    "stroke-width": "2"
-  },
-  "russia": {
-    "fill": "#008000",
-    "stroke": "black",
-    "stroke-width": "2"
-  },
-  "spain": {
-    "fill": "#FF8080",
-    "stroke": "black",
-    "stroke-width": "2"
-  },
-  "turkey": {
-    "fill": "#FFFF00",
-    "stroke": "black",
-    "stroke-width": "2"
-  },
-  "ukraine": {
-    "fill": "#FF00FF",
-    "stroke": "black",
-    "stroke-width": "2"
-  },
-  "unitbritain": {
-    "fill": "deepskyblue",
-    "stroke": "black",
-    "fill-opacity": "0.90"
-  },
-  "unitegypt": {
-    "fill": "#808000",
-    "stroke": "black",
-    "fill-opacity": "0.90"
-  },
-  "unitfrance": {
-    "fill": "#00FFFF",
-    "stroke": "black",
-    "fill-opacity": "0.90"
-  },
-  "unitgermany": {
-    "fill": "darkgrey",
-    "stroke": "black",
-    "fill-opacity": "0.90"
-  },
-  "unititaly": {
-    "fill": "#80FF80",
-    "stroke": "black",
-    "fill-opacity": "0.90"
-  },
-  "unitpoland": {
-    "fill": "#FF0000",
-    "stroke": "black",
-    "fill-opacity": "0.90"
-  },
-  "unitrussia": {
-    "fill": "#008000",
-    "stroke": "black",
-    "fill-opacity": "0.90"
-  },
-  "unitspain": {
-    "fill": "#FF8080",
-    "stroke": "black",
-    "fill-opacity": "0.90"
-  },
-  "unitturkey": {
-    "fill": "#FFFF00",
-    "stroke": "black",
-    "fill-opacity": "0.90"
-  },
-  "unitukraine": {
-    "fill": "#FF00FF",
-    "stroke": "black",
-    "fill-opacity": "0.90"
-  },
-  "supportorder": {
-    "stroke-width": "2",
-    "fill": "none",
-    "stroke-dasharray": "5,5"
-  },
-  "convoyorder": {
-    "stroke-dasharray": "15,5",
-    "stroke-width": "2",
-    "fill": "none"
-  },
-  "shadowdash": {
-    "stroke-width": "4",
-    "fill": "none",
-    "stroke": "black",
-    "opacity": "0.45"
-  },
-  "varwidthorder": {
-    "fill": "none"
-  },
-  "varwidthshadow": {
-    "fill": "none",
-    "stroke": "black"
-  },
-  "style1": {
-    "fill": "darkGray"
   }
 }
